Added keys for Comments and Posts, some fixes
