+++ src/App.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import ButtonPosts from './components/ButtonPosts'
 import PostList from './components/PostList'
+import LoadData from './components/FetchData'
 
 import './App.css' 
 
@@ -18,19 +19,7 @@ class App extends React.Component {
       isLoading: true,
     })
 
-    const responsePosts = await fetch('https://jsonplaceholder.typicode.com/posts');
-    const responseUsers = await fetch('https://jsonplaceholder.typicode.com/users');
-    const responseComments = await fetch('https://jsonplaceholder.typicode.com/comments');
-
-    const posts = await responsePosts.json();
-    const users = await responseUsers.json();
-    const comments = await responseComments.json();
-
-    const postsWithUser = posts.map(post => ({
-      ...post,
-      user: users.find(user => user.id === post.userId),
-      comments: comments.filter(comment => post.id === comment.postId),
-    }));
+    const postsWithUser = await LoadData();
 
     this.setState({
       currentPosts: postsWithUser,
@@ -44,7 +33,7 @@ class App extends React.Component {
     const {value} = e.target;
     this.setState(prevState => ({
       inputValue: value,
-      filteredPosts: [...prevState.currentPosts].filter(post=> post.title.includes(value) || post.body.includes(value))
+      filteredPosts: prevState.currentPosts.filter(post=> post.title.includes(value) || post.body.includes(value))
       }))
   }
