+++ src/components/Comment.jsx
@@ -1,6 +1,6 @@
 import React from 'react'
 
-const Comment = ({currentComment}) => (
+const Comment = ({ currentComment}) => (
   <div>
     <div className='comment__author'>{currentComment.email}</div>
     <div className='comment__body'>{currentComment.body}</div>
